+++ activerecord/lib/arel/nodes/delete_statement.rb
@@ -3,8 +3,7 @@
 module Arel # :nodoc: all
   module Nodes
     class DeleteStatement < Arel::Nodes::Node
-      attr_accessor :left, :right
-      attr_accessor :limit
+      attr_accessor :left, :right, :orders, :limit, :key
 
       alias :relation :left
       alias :relation= :left=
@@ -15,6 +14,9 @@ module Arel # :nodoc: all
         super()
         @left = relation
         @right = wheres
+        @orders = []
+        @limit = nil
+        @key = nil
       end
 
       def initialize_copy(other)
@@ -24,13 +26,16 @@ module Arel # :nodoc: all
       end
 
       def hash
-        [self.class, @left, @right].hash
+        [self.class, @left, @right, @orders, @limit, @key].hash
       end
 
       def eql?(other)
         self.class == other.class &&
           self.left == other.left &&
-          self.right == other.right
+          self.right == other.right &&
+          self.orders == other.orders &&
+          self.limit == other.limit &&
+          self.key == other.key
       end
       alias :== :eql?
     end
